Finish sales order event and status update API

+++ ScmssApiServer/DomainServices/SalesOrdersService.cs
@@ -19,32 +19,64 @@ namespace ScmssApiServer.DomainServices
             _mapper = mapper;
         }
 
-        public async Task<SalesOrderEvent> AddManualEvent(int orderId, OrderEventInputDto dto)
+        public async Task<OrderEventDto> AddEvent(int orderId, OrderEventCreateDto dto)
         {
-            var item = _mapper.Map<SalesOrderEvent>(dto);
-            SalesOrder order = await GetSalesOrderOrThrowAsync(orderId);
-            order.AddManualEvent(dto.Type, dto.Location, dto.Message);
+            SalesOrder? order = await _dbContext.SalesOrders.FirstOrDefaultAsync(i => i.Id == orderId);
+            if (order == null)
+            {
+                throw new EntityNotFoundException();
+            }
+
+            SalesOrderEvent item = order.AddManualEvent(dto.Type, dto.Location, dto.Message);
+
             await _dbContext.SaveChangesAsync();
-            return item;
+            return GetOrderEventDto(item);
         }
 
         public async Task<SalesOrderDto> CreateSalesOrderAsync(SalesOrderCreateDto dto, string userId)
         {
-            var item = _mapper.Map<SalesOrder>(dto);
+            Customer? customer = await _dbContext.Customers.FindAsync(dto.CustomerId);
+            if (customer == null)
+            {
+                throw new EntityNotFoundException("Customer not found");
+            }
+
+            var item = new SalesOrder
+            {
+                ToLocation = dto.ToLocation ?? customer.DefaultLocation,
+                CustomerId = dto.CustomerId,
+                CreateUserId = userId,
+            };
+
+            if (dto.ProductionFacilityId != null)
+            {
+                int facilityId = (int)dto.ProductionFacilityId;
+                item.ProductionFacilityId = facilityId;
+                item.FromLocation = await GetLocationOfProductionFacilityAsync(facilityId);
+            }
+
             await AddOrderItemsFromDtos(item, dto.Items);
             item.Start(userId);
+
             _dbContext.SalesOrders.Add(item);
             await _dbContext.SaveChangesAsync();
             return GetSalesOrderDto(item);
         }
 
-        public async Task<SalesOrderEvent> EditEvent(int id, int orderId, OrderEventInputDto dto)
+        public async Task<OrderEventDto> UpdateEvent(int id, int orderId, OrderEventUpdateDto dto)
         {
-            var item = _mapper.Map<SalesOrderEvent>(dto);
-            SalesOrder order = await GetSalesOrderOrThrowAsync(id);
-            order.EditEvent(id, dto.Location, dto.Message);
+            SalesOrder? order = await _dbContext.SalesOrders
+                .Include(i => i.Events)
+                .FirstOrDefaultAsync(i => i.Id == orderId);
+            if (order == null)
+            {
+                throw new EntityNotFoundException();
+            }
+
+            SalesOrderEvent item = order.UpdateEvent(id, dto.Location, dto.Message);
+
             await _dbContext.SaveChangesAsync();
-            return item;
+            return GetOrderEventDto(item);
         }
 
         public async Task<SalesOrderDto?> GetSalesOrderAsync(int id)
@@ -64,44 +96,82 @@ namespace ScmssApiServer.DomainServices
             IList<SalesOrder> items = await _dbContext.SalesOrders
                 .Include(i => i.CreateUser)
                 .Include(i => i.FinishUser)
-                .Include(i => i.Events)
                 .Include(i => i.ProductionFacility)
-                .Include(i => i.Items)
                 .ToListAsync();
             return _mapper.Map<IList<SalesOrderDto>>(items);
         }
 
-        public async Task<SalesOrderDto> UpdateSalesOrderAsync(int id, SalesOrderInputDto dto, string userId)
+        public async Task<SalesOrderDto> UpdateSalesOrderAsync(int id,
+                                                               SalesOrderUpdateDto dto,
+                                                               string userId)
         {
             SalesOrder? item = await _dbContext.SalesOrders
                 .Include(i => i.Items)
                 .Include(i => i.Events)
+                .Include(i => i.Customer)
                 .FirstOrDefaultAsync(i => i.Id == id);
             if (item == null)
             {
                 throw new EntityNotFoundException();
             }
 
-            _mapper.Map(dto, item);
-            switch (dto.Status)
-            {
-                case SalesOrderStatusSelection.Completed:
-                    item.Complete(userId);
-                    break;
-                case SalesOrderStatusSelection.Canceled:
-                    item.Cancel(userId);
-                    break;
-                case SalesOrderStatusSelection.Returned:
-                    item.Return(userId);
-                    break;
-            }
             if (dto.PaymentCompleted ?? false)
             {
                 item.CompletePayment();
+                await _dbContext.SaveChangesAsync();
+                return GetSalesOrderDto(item);
             }
 
-            _dbContext.RemoveRange(item.Items);
-            await AddOrderItemsFromDtos(item, dto.Items);
+            if (item.Finished)
+            {
+                throw new InvalidDomainOperationException(
+                        "Cannot update order because it is finished"
+                    );
+            }
+
+            if (item.Status == OrderStatus.Processing)
+            {
+                if (dto.ProductionFacilityId != null)
+                {
+                    int facilityId = (int)dto.ProductionFacilityId;
+                    item.ProductionFacilityId = facilityId;
+                    item.FromLocation = await GetLocationOfProductionFacilityAsync(facilityId);
+                }
+
+                if (dto.Items != null)
+                {
+                    _dbContext.RemoveRange(item.Items);
+                    await AddOrderItemsFromDtos(item, dto.Items);
+                }
+            }
+
+            if (item.Status != OrderStatus.Delivered)
+            {
+                item.ToLocation = dto.ToLocation ?? item.Customer.DefaultLocation;
+            }
+
+            switch (dto.Status)
+            {
+                case OrderStatusSelection.Delivering:
+                    item.StartDelivery();
+                    break;
+
+                case OrderStatusSelection.Delivered:
+                    item.FinishDelivery();
+                    break;
+
+                case OrderStatusSelection.Completed:
+                    item.Complete(userId);
+                    break;
+
+                case OrderStatusSelection.Canceled:
+                    item.Cancel(userId);
+                    break;
+
+                case OrderStatusSelection.Returned:
+                    item.Return(userId);
+                    break;
+            }
 
             await _dbContext.SaveChangesAsync();
             return GetSalesOrderDto(item);
@@ -112,14 +182,21 @@ namespace ScmssApiServer.DomainServices
             return _mapper.Map<SalesOrderDto>(item);
         }
 
-        private async Task<SalesOrder> GetSalesOrderOrThrowAsync(int id)
+        private OrderEventDto GetOrderEventDto(SalesOrderEvent item)
         {
-            SalesOrder? item = await _dbContext.SalesOrders.FindAsync(id);
-            if (item == null)
+            return _mapper.Map<OrderEventDto>(item);
+        }
+
+        private async Task<string> GetLocationOfProductionFacilityAsync(int facilityId)
+        {
+            ProductionFacility? facility = await _dbContext
+                .ProductionFacilities
+                .FindAsync(facilityId);
+            if (facility == null)
             {
-                throw new EntityNotFoundException();
+                throw new EntityNotFoundException("Production facility not found");
             }
-            return item;
+            return facility.Location;
         }
 
         private async Task AddOrderItemsFromDtos(SalesOrder order, IEnumerable<OrderItemInputDto> dtos)
